# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "wynpay",
        "title": "Gin-Server Swagger API接口文档",
        "contact": {},
        "version": "v0.0.1"
    },
    "paths": {
        "/api/formation/v1/save": {
            "post": {
                "description": "保存连接信息并生成连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SaveFormation 保存连接信息并生成连接"
                ],
                "summary": "SaveFormation",
                "parameters": [
                    {
                        "description": "phone,password",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FormationMode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResData"
                        }
                    }
                }
            }
        },
        "/api/user/v1/login": {
            "post": {
                "description": "帐号登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login 帐号登陆"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "phone,password",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.FormationMode": {
            "type": "object",
            "properties": {
                "currency_code": {
                    "type": "string"
                },
                "utm_campaign": {
                    "type": "string"
                },
                "utm_medium": {
                    "type": "string"
                },
                "utm_source": {
                    "type": "string"
                }
            }
        },
        "request.UserLogin": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "response.ResData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}